Updated HUD & Added VIllager Palettes
The HUD now includes a Build Menu button that accesses the Build Menu. HUD buttons now also have the Hotkey Bind script, but it seems to be intercepted by the Game Manager or some other script as it does not work and throws errors. Created and added all possible color palettes for villagers, however the shader itself is not fully complete.